added header to layout
