+++ src/components/Header.scss
@@ -1,0 +1,13 @@
+.header {
+  height: 100px;
+  background-color: white;
+  display: flex;
+  align-items: center;
+  padding-left: 10px;
+  padding-right: 10px;
+
+  img {
+    height: 80px;
+    margin: 0;
+  }
+}
