+++ src/components/layout.jsx
@@ -1,40 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { StaticQuery, graphql } from 'gatsby';
 
-import Header from './header';
+import Header from './Header';
 import './layout.css';
 
 const Layout = ({ children }) => (
-  <StaticQuery
-    query={graphql`
-      query SiteTitleQuery {
-        site {
-          siteMetadata {
-            title
-          }
-        }
-      }
-    `}
-    render={data => (
-      <React.Fragment>
-        <Header siteTitle={data.site.siteMetadata.title} />
-        <div
-          style={{
-            margin: '0 auto',
-            maxWidth: 960,
-            padding: '0px 1.0875rem 1.45rem',
-            paddingTop: 0,
-          }}
-        >
-          {children}
-          <footer>
-            © {new Date().getFullYear()}, Built with {' '} <a href="https://www.gatsbyjs.org">Gatsby</a>
-          </footer>
-        </div>
-      </React.Fragment>
-    )}
-  />
+  <React.Fragment>
+    <Header />
+    {children}
+  </React.Fragment>
 );
 
 Layout.propTypes = {
