+++ src/pages/index.jsx
@@ -4,13 +4,17 @@ import Image from '../components/Image';
 
 import './index.scss';
 
-//import Layout from '../components/layout';
+import Layout from '../components/layout';
 // import SEO from '../components/seo';
 
 const IndexPage = () => (
-  <div className="main">
-    <div className="logoPlaceholder"><Image /></div>
-  </div>
+  <Layout>
+    <div className="main">
+      <div className="logoPlaceholder">
+        <Image />
+      </div>
+    </div>
+  </Layout>
 );
 
 export default IndexPage;
